requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
psycopg2-binary>=2.9
requests>=2.31
django-cors-headers>=4.4
drf-spectacular>=0.27,<1.0
PyPDF2>=3.0
python-docx>=1.1
